+++ R/bertin.r
@@ -335,7 +335,7 @@ bertinBase <- function(nrow, ncol, labels = "", labels.elements = "",
   }
 
   colorRow <- function(cr) {
-    par(new = TRUE) # next plot will overplot not earse the old one, necessary for setting the same regions
+    old_par <- par(new = TRUE) # next plot will overplot not erase the old one, necessary for setting the same regions
     plot.new()
     # plot.window(xlim=0:1, ylim=0:1) #, xaxs="i", yaxs="i")#, asp =nrow/ncol)
     if (cr >= 1 & cr <= nrow) { # color current row cr
@@ -348,10 +348,11 @@ bertinBase <- function(nrow, ncol, labels = "", labels.elements = "",
         labels = labels.rows, col = col.mark.text, cex = cex.text
       )
     }
+    par(old_par)
   }
 
   colorColumn <- function(cc) {
-    par(new = TRUE) # next plot will overplot not earse the old one, necessary for setting the same regions
+    old_par <- par(new = TRUE) # next plot will overplot not earse the old one, necessary for setting the same regions
     plot.new()
     # plot.window(xlim=0:1, ylim=0:1) #, xaxs="i", yaxs="i")#, asp =nrow/ncol)
     if (cc >= 1 & cc <= ncol) { # color current column cc
@@ -373,6 +374,7 @@ bertinBase <- function(nrow, ncol, labels = "", labels.elements = "",
       segments(x.lines[cc], y1.lines, x.lines[cc], y2.lines[cc], lwd = 3, col = "white") # overplot old stroke in white
       segments(x.lines[cc], y1.lines, x.lines[cc], y2.lines[cc], col = col.mark.fill)
     }
+    par(old_par)
   }
 
   renewColumn <- function(cc) {
@@ -440,7 +442,7 @@ bertinBase <- function(nrow, ncol, labels = "", labels.elements = "",
   # set plotting parameters
   # old.par <- par(no.readonly = TRUE)    # save parameters
   # on.exit(par(old.par))                 # reset old par when done
-  op <- par(oma = rep(0, 4), mar = rep(0, 4), xaxs = "i", yaxs = "i")
+  old_par <- par(oma = rep(0, 4), mar = rep(0, 4), xaxs = "i", yaxs = "i")
   if (print) { # in case no new printing should occur
     par(new = FALSE)
   } else {
@@ -466,7 +468,7 @@ bertinBase <- function(nrow, ncol, labels = "", labels.elements = "",
     colorRow(cr)
     colorColumn(cc)
   }
-  # par(op)
+  par(old_par)
   invisible(NULL)
 }
 
@@ -786,16 +788,14 @@ bertinCluster <- function(x, dmethod = c("euclidean", "euclidean"),
 
   x <- x[con.ord, el.ord] # reorder repgrid object
 
+  old_par <- par(fig = c(xsegs[c(1, 4)], ysegs[c(2, 4)]))
   plot.new()
-  par(fig = c(xsegs[c(1, 4)], ysegs[c(2, 4)]), new = TRUE)
-  # par(fig = c(0, .8, .2, 1), new=T)
-
-  bertin(x, xlim = xlim.bertin, ylim = ylim.bertin, add = TRUE, ...) # print reordered bertin
+  bertin(x, xlim = xlim.bertin, ylim = ylim.bertin, add = FALSE, ...) # print reordered bertin
 
   # x dendrogram (horizontal) elements
   if (!is.na(cmethod[2])) {
     dend.x.fig <- c(xsegs[2:3], ysegs[1:2]) + c(0, 0, y.off, -y.off) # adjust for offsets
-    par(fig = dend.x.fig, new = T, mar = c(0, 0, 0, 0))
+    par(fig = dend.x.fig, new = TRUE, mar = c(0, 0, 0, 0)) # trick: new = TRUE avoids opening a new plot window in next plot
     ymax.el <- attr(dend.el, "height")
     plot(dend.el,
       horiz = F, xlab = "", xaxs = "i", yaxs = "i", yaxt = "n",
@@ -819,17 +819,7 @@ bertinCluster <- function(x, dmethod = c("euclidean", "euclidean"),
       axis(1, las = 1, cex.axis = cex.axis, col = col.axis, col.axis = col.axis)
     }
   }
+  par(old_par)
   # return hclust objects for elements and constructs
   invisible(list(constructs = fit.constructs, elements = fit.elements))
 }
-
-# TODO: use of layout does not work with bertinCluster
-# a future version could use layout
-# layout (matrix(1:4), 2)
-# bertinCluster(bell2010)
-
-# bertinCluster(bell2010, type="t", bor=grey(.5))
-# dev.new()
-# bertinCluster(bell2010, type="t", dm="manhattan", cm="single")
-# dev.new()
-# bertinCluster(bell2010, type="t", dm="manhattan", cm="centroid")