+++ R/repgrid-basicops.r
@@ -96,6 +96,13 @@ stop_if_not_integerish <- function(x, arg = NULL) {
 }
 
 
+stop_if_not_inherits <- function(x, what, argname = "x") {
+  if (!inherits(x, what)) {
+    stop("`", argname, "` must have class <", what, "> not <", class(x)[1], ">", call. = FALSE)
+  }
+}
+
+
 ############################# EXTRACT AND SET #################################
 
 ## S4 methods
@@ -510,9 +517,9 @@ setScale <- function(x, min, max, step, ...) { # ... needes for makeRepgrid call
     }
     x@scale$max <- max
   }
-  if (!missing(step)) {
-    x@scale$step <- step
-  }
+  # if (!missing(step)) {
+  #   x@scale$step <- step
+  # }
   x
 }
 # setScale(x, min=1, max=5, step=1)
@@ -1451,6 +1458,7 @@ makeRepgrid <- function(args) {
   x <- do.call(rg.setCoupled, l) # if no coupled argument then coupled=TRUE
   l <- c(list(x = x), args) # make a new repgrid object
   x <- do.call(setScale, l) # set scale if min and max arg is provided
+  preferredPoles(x) <- l$preferred_pole
   x
 }
 # args <- list( name=c("element_1", "element_2", "element_3", "element_4"),